+++ packages/kit/src/views/Market/MarketDetailV2/components/InformationTabs/components/Portfolio/hooks/usePortfolioData.ts
@@ -8,12 +8,14 @@ interface IUsePortfolioDataProps {
   tokenAddress: string;
   networkId: string;
   accountAddress?: string;
+  xpub?: string;
 }
 
 export function usePortfolioData({
   tokenAddress,
   networkId,
   accountAddress,
+  xpub,
 }: IUsePortfolioDataProps) {
   const {
     result: portfolioData,
@@ -29,9 +31,10 @@ export function usePortfolioData({
         tokenAddress,
         networkId,
         accountAddress,
+        xpub,
       });
     },
-    [tokenAddress, networkId, accountAddress],
+    [tokenAddress, networkId, accountAddress, xpub],
     {
       watchLoading: true,
       pollingInterval: timerUtils.getTimeDurationMs({ seconds: 5 }),
