+++ packages/kit/src/views/Market/MarketDetailV2/layouts/DesktopLayout.tsx
@@ -17,7 +17,7 @@ import {
   TokenSupplementaryInfo,
 } from '../components';
 import { usePortfolioData } from '../components/InformationTabs/components/Portfolio/hooks/usePortfolioData';
-import { useNetworkAccountAddress } from '../components/InformationTabs/hooks/useNetworkAccountAddress';
+import { useNetworkAccount } from '../components/InformationTabs/hooks/useNetworkAccount';
 import { DesktopInformationTabs } from '../components/InformationTabs/layout/DesktopInformationTabs';
 import { useTokenDetail } from '../hooks/useTokenDetail';
 
@@ -25,12 +25,13 @@ export function DesktopLayout() {
   const { tokenAddress, networkId, tokenDetail, isNative, websocketConfig } =
     useTokenDetail();
 
-  const { accountAddress } = useNetworkAccountAddress(networkId);
+  const { accountAddress, xpub } = useNetworkAccount(networkId);
 
   const { portfolioData, isRefreshing } = usePortfolioData({
     tokenAddress,
     networkId,
     accountAddress,
+    xpub,
   });
 
   const isBTCNetwork = networkUtils.isBTCNetwork(networkId);
